Source Commit: Continueing with the refactor, I changed a few variables from floats to ints. Several blueprints that used these variables needed to be editted to use the new variables instead of the old.

+++ Source/BoomerShooter/Gun.h
@@ -94,10 +94,10 @@ class BOOMERSHOOTER_API AGun : public AWeapon
 		TSubclassOf<class AProjectile> ProjectileClass;
 
 		UPROPERTY(EditAnywhere, Category = "Combat")
-		float ModManaCost = 0.f;
+		int ModManaCost = 0;
 
 		UPROPERTY(EditAnywhere, Category = "Combat")
-		float BaseManaCost = 0.f;
+		int BaseManaCost = 0;
 		
 		class ABaseCharacter* PlayerCharacter;
